Dialogue comes up with E button
The dialogue for an NPC is now signified with a sprite above their head. When the player presses e they can initiate dialogue.
