--- Basilisk-Zero/Assets/ArtTemplate/UI/ui x1.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c79eca8d4ac60fcc1973883f65e7e842f298f5c2b5d4b499301b951317be26c8
-size 2693
