+++ Basilisk-Zero/Assets/Scripts/Dialogue/YarnInteractable.cs.meta
